Refactor and abstract media player events/tracking into seperate module to be more DRY
